backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1

# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# File Processing
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2

# Server (production)
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
